+++ packages/web/src/javascripts/Components/AccountMenu/User.tsx
@@ -1,6 +1,7 @@
 import { observer } from 'mobx-react-lite'
 import { User as UserType } from '@standardnotes/snjs'
 import { useApplication } from '../ApplicationProvider'
+import { c } from 'ttag'
 
 const User = () => {
   const application = useApplication()
@@ -12,9 +13,10 @@ const User = () => {
     <div className="sk-panel-section">
       {application.syncStatusController.errorMessage && (
         <div className="sk-notification danger">
-          <div className="sk-notification-title">Sync Unreachable</div>
+          <div className="sk-notification-title">{c('Title').t`Sync Unreachable`}</div>
           <div className="sk-notification-text">
-            Hmm...we can't seem to sync your account. The reason: {application.syncStatusController.errorMessage}
+            {c('Error')
+              .t`Hmm...we can't seem to sync your account. The reason: ${application.syncStatusController.errorMessage}`}
           </div>
           <a
             className="sk-a info-contrast sk-bold sk-panel-row"
@@ -22,7 +24,7 @@ const User = () => {
             rel="noopener"
             target="_blank"
           >
-            Need help?
+            {c('Action').t`Need help?`}
           </a>
         </div>
       )}