+++ packages/web/src/javascripts/Components/ConfirmDeleteAccountModal/ConfirmDeleteAccountModal.tsx
@@ -1,5 +1,6 @@
 import { observer } from 'mobx-react-lite'
 import { useCallback, useRef } from 'react'
+import { c } from 'ttag'
 import { STRING_DELETE_ACCOUNT_CONFIRMATION } from '@/Constants/Strings'
 import Button from '@/Components/Button/Button'
 import { WebApplication } from '@/Application/WebApplication'
@@ -25,22 +26,22 @@ const ConfirmDeleteAccountModal = ({ application }: Props) => {
   return (
     <AlertDialog closeDialog={closeDialog}>
       <div className="flex items-center justify-between text-lg font-bold">
-        Delete account?
+        {c('Title').t`Delete account?`}
         <button className="rounded p-1 font-bold hover:bg-contrast" onClick={closeDialog}>
           <Icon type="close" />
         </button>
       </div>
       <div className="sk-panel-row">
         <div>
-          <p className="text-base text-foreground lg:text-sm">{STRING_DELETE_ACCOUNT_CONFIRMATION}</p>
+          <p className="text-base text-foreground lg:text-sm">{STRING_DELETE_ACCOUNT_CONFIRMATION()}</p>
         </div>
       </div>
       <div className="mt-4 flex justify-end gap-2">
         <Button ref={cancelRef} onClick={closeDialog}>
-          Cancel
+          {c('Action').t`Cancel`}
         </Button>
         <Button primary colorStyle="danger" onClick={confirm}>
-          Delete my account for good
+          {c('Action').t`Delete my account for good`}
         </Button>
       </div>
     </AlertDialog>