+++ packages/web/src/javascripts/Components/NoSubscriptionBanner/NoSubscriptionBanner.tsx
@@ -30,9 +30,11 @@ const NoSubscriptionBanner = ({
         <h1 className="sk-h3 m-0 text-sm font-semibold">{title}</h1>
       </div>
       <p className="col-start-1 col-end-3 m-0 mt-1 text-sm">{message}</p>
-      <Button primary small className="col-start-1 col-end-3 mt-3 justify-self-start uppercase" onClick={onClick}>
-        Upgrade Features
-      </Button>
+      {application.canShowPurchaseFlow() && (
+        <Button primary small className="col-start-1 col-end-3 mt-3 justify-self-start uppercase" onClick={onClick}>
+          Upgrade Features
+        </Button>
+      )}
     </div>
   )
 }