+++ packages/web/src/javascripts/Components/Preferences/Panes/Account/SignOutView.tsx
@@ -8,6 +8,7 @@ import { Subtitle, Title, Text } from '../../PreferencesComponents/Content'
 import PreferencesGroup from '../../PreferencesComponents/PreferencesGroup'
 import PreferencesSegment from '../../PreferencesComponents/PreferencesSegment'
 import ClearSessionDataView from './ClearSessionDataView'
+import { c } from 'ttag'
 
 type Props = {
   application: WebApplication
@@ -18,27 +19,27 @@ const SignOutView: FunctionComponent<Props> = observer(({ application }) => {
     <>
       <PreferencesGroup>
         <PreferencesSegment>
-          <Title>Sign out</Title>
-          <Subtitle>Other devices</Subtitle>
-          <Text>Want to sign out on all devices except this one?</Text>
+          <Title>{c('Title').t`Sign out`}</Title>
+          <Subtitle>{c('Subtitle').t`Other devices`}</Subtitle>
+          <Text>{c('Info').t`Want to sign out on all devices except this one?`}</Text>
           <div className="mt-3 flex flex-row flex-wrap gap-3">
             <Button
-              label="Sign out other sessions"
+              label={c('Action').t`Sign out other sessions`}
               onClick={() => {
                 application.accountMenuController.setOtherSessionsSignOut(true)
               }}
             />
-            <Button label="Manage sessions" onClick={() => application.openSessionsModal()} />
+            <Button label={c('Action').t`Manage sessions`} onClick={() => application.openSessionsModal()} />
           </div>
         </PreferencesSegment>
         <HorizontalSeparator classes="my-4" />
         <PreferencesSegment>
-          <Subtitle>This workspace</Subtitle>
-          <Text>Remove all data related to the current workspace from the application.</Text>
+          <Subtitle>{c('Subtitle').t`This workspace`}</Subtitle>
+          <Text>{c('Info').t`Remove all data related to the current workspace from the application.`}</Text>
           <Button
             className="mt-3"
             colorStyle="danger"
-            label="Sign out workspace"
+            label={c('Action').t`Sign out workspace`}
             onClick={() => {
               application.accountMenuController.setSigningOut(true)
             }}