+++ packages/web/src/javascripts/Event/ApplicationEventObserver.ts
@@ -47,7 +47,9 @@ export class ApplicationEventObserver implements EventObserverInterface {
           const route = this.routeService.getRoute()
           switch (route.type) {
             case RouteType.Purchase:
-              void this.purchaseFlowController.openPurchaseFlow()
+              if (this.application.canShowPurchaseFlow()) {
+                void this.purchaseFlowController.openPurchaseFlow()
+              }
 
               break
             case RouteType.Settings: {