+++ packages/clipper/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.588](https://github.com/standardnotes/app/compare/@standardnotes/clipper@1.1.587...@standardnotes/clipper@1.1.588) (2026-06-05)
+
+**Note:** Version bump only for package @standardnotes/clipper
+
 ## [1.1.587](https://github.com/standardnotes/app/compare/@standardnotes/clipper@1.1.586...@standardnotes/clipper@1.1.587) (2026-05-27)
 
 **Note:** Version bump only for package @standardnotes/clipper